+++ src/assets/styles/userWrite.module.scss
@@ -1,29 +1,35 @@
 .createPostContainer {
-    max-width: 600px;
-    margin: 0 auto;
-    padding: 20px;
-    background-color: #f9f9f9;
-    border-radius: 10px;
-    box-shadow: 0 0 10px rgba(0, 0, 0, 0.1);
-  }
-  
-  .createPostContainer h2 {
-    margin-bottom: 20px;
-  }
-  
-  .createPostContainer .form-group {
-    margin-bottom: 15px;
-  }
-  
-  .createPostContainer .form-control {
-    border-radius: 5px;
-  }
-  
-  .createPostContainer .btn-primary {
-    background-color: #007bff;
-    border-color: #007bff;
-    border-radius: 5px;
-    padding: 10px 20px;
-    font-size: 16px;
-  }
-  
+  max-width: 600px;
+  margin: 0 auto;
+  padding: 20px;
+  background-color: #f9f9f9;
+  border-radius: 10px;
+  box-shadow: 0 0 10px rgba(0, 0, 0, 0.1);
+}
+
+.createPostContainer h2 {
+  margin-bottom: 20px;
+}
+
+.createPostContainer .form-group {
+  margin-bottom: 15px;
+}
+
+.createPostContainer .form-control {
+  border-radius: 5px;
+}
+
+.createPostContainer .btn-primary {
+  background-color: #007bff;
+  border-color: #007bff;
+  border-radius: 5px;
+  padding: 10px 20px;
+  font-size: 16px;
+}
+
+.tableCellCenter {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  gap: 10px; /* 버튼 간 간격을 조정합니다 */
+}
