+++ src/pages/Community/Community.js
@@ -1,10 +1,10 @@
 import React, { useEffect, useState } from 'react';
 import { Link } from 'react-router-dom';
-import Loading from '../components/common/Loading';
+import Loading from '../../components/common/Loading';
 // import api from '../services/api';
 import Table from 'react-bootstrap/Table';
-import CommonPagination from '../components/common/CommonPagination';
-import styles from '../assets/styles/community/community.module.scss';
+import CommonPagination from '../../components/common/CommonPagination';
+import styles from '../../assets/styles/community/community.module.scss';
 import axios from 'axios';
 
 const Community = () => {
