+++ src/pages/Mypage/UserInfo.js
@@ -1,6 +1,6 @@
 import React, { useEffect, useState } from 'react';
-import Loading from '../components/common/Loading';
-import api from '../services/api';
+import Loading from '../../components/common/Loading';
+import api from '../../services/api';
 
 const UserInfo = ({ onEditClick }) => {
   const [userInfo, setUserInfo] = useState(null);
