+++ src/pages/Mypage/UserUpdate.js
@@ -2,7 +2,7 @@ import React, { useEffect } from 'react';
 import { useForm } from 'react-hook-form';
 import { yupResolver } from '@hookform/resolvers/yup';
 import * as yup from 'yup';
-import api from '../services/api'
+import api from '../../services/api'
 
 // 유효성 검사 스키마 정의
 const schema = yup.object().shape({
