+++ src/pages/Mypage/UserUpdateTest.js
@@ -1,7 +1,7 @@
 import React, { useEffect, useState } from 'react';
-import api from '../services/api';
+import api from '../../services/api';
 import { useNavigate } from 'react-router-dom';
-import Loading from '../components/common/Loading';
+import Loading from '../../components/common/Loading';
 import Button from 'react-bootstrap/Button';
 import Form from 'react-bootstrap/Form';
 import * as yup from 'yup';
